Optimize

+++ Runtime/Shaders/EasyPBR_Input.hlsl
@@ -115,10 +115,10 @@ CBUFFER_START(UnityPerMaterial)
     half _GlitterScale;
     half _GlitterSize;
     half _GlitterTilt;
-    float _GlitterSparsity;
-    float _GlitterIridescence;
-    float _GlitterIridescenceShift;
-    float _GlitterBaseReflection;
+    half _GlitterSparsity;
+    half _GlitterIridescence;
+    half _GlitterIridescenceShift;
+    half _GlitterBaseReflection;
 
     half4 _SSSColor;
     half _SSSIntensity;
